Split CLI into slumber_cli crate
