+++ crates/slumber_cli/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+authors = ["Lucas Pickering <lucas@lucaspickering.me>"]
+description = "Command line interface for Slumber. Not intended for external use."
+edition = "2021"
+homepage = "https://slumber.lucaspickering.me"
+keywords = ["rest", "http", "terminal", "tui"]
+license = "MIT"
+name = "slumber_cli"
+repository = "https://github.com/LucasPickering/slumber"
+version = "1.7.0"
+# Rely on parent for rust-version
+
+[dependencies]
+anyhow = "^1.0.0"
+bytesize = {version = "1.3.0", default-features = false}
+clap = {version = "^4.4.2", features = ["derive"]}
+dialoguer = {version = "^0.11.0", default-features = false, features = ["password"]}
+indexmap = {version = "^2.0.0", features = ["serde"]}
+itertools = "^0.12.0"
+reqwest = {version = "^0.12.0", default-features = false}
+serde = "^1.0.0"
+serde_yaml = {version = "^0.9.0", default-features = false}
+slumber_core = {path = "../slumber_core", version = "1.7.0"}
+tracing = "^0.1.37"
