+++ crates/slumber_cli/src/commands.rs
@@ -1,0 +1,6 @@
+pub mod collections;
+pub mod generate;
+pub mod history;
+pub mod import;
+pub mod request;
+pub mod show;
